pyproject.toml
[tool.poetry]
name = "sptools"
version = "0.1.0"
description = ""
authors = ["Charles Bine <charles.bine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
spotipy = "^2.23.0"
click = "^8.1.3"
python-dotenv = "^1.0.0"
tinydb = "^4.7.1"
tomli = "^2.0.1"
loguru = "^0.6.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
